+++ praw/models/reddit/submission.py
@@ -1,7 +1,7 @@
 """Provide the Submission class."""
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 from urllib.parse import urljoin
-from logging import getLogger
+from warnings import warn
 
 from prawcore import Conflict
 
@@ -19,8 +19,6 @@ from .subreddit import Subreddit
 
 if TYPE_CHECKING:  # pragma: no cover
     import praw
-
-logger = getLogger("praw")
 
 
 class SubmissionFlair:
@@ -594,8 +592,13 @@ class Submission(SubmissionListingMixin, UserContentMixin, FullnameMixin, Reddit
             value = Subreddit(self._reddit, value)
         elif attribute == "poll_data":
             value = PollData(self._reddit, value)
-        elif attribute == "comment_sort" and self._fetched:
-            logger.warning(
+        elif (
+            self._reddit.config.warn_comment_sort
+            and attribute == "comment_sort"
+            and hasattr(self, "_fetched")
+            and self._fetched
+        ):
+            warn(
                 "The comments for this submission have already been fetched, "
                 "so the updated comment_sort will not have any effect"
             )